Bug-fix 521: Check that new passwords match the required criteria
When changing a user's password, the app asks for a new password with at least one letter and one number. However, this was not enforced.
Now, the changeSettingController checks the new password when the Update button is clicked, and refuses it if it isn't valid.

The changeSettingController used to require the user's old password to be at least 5 characters long when updating the password, or their old email to be at least 4 characters long when updating the email. This was removed.