Merge-fix: clean up code (style, comments)
I added comments to identify sections of the code written by Eason and to clarify some lines of code. I also commented out the printouts to the console used for testing. The function opened(e) from educationalMaterialController was unexplained and unused, so I removed it. In addition, small modifications to the style were made (naming, spacing, etc.)