Bug-fix: scroll to bottom not working
'Scroll to bottom' logging was not working most of the time, because the screen position that was used to check whether the user scrolled to the bottom included a decimal fraction. This decimal number was then compared to an integer, which would rarely output true (example: if (512==512.3333) log 'scrolled to bottom').

I rearranged the calculation to equal close to 0 when the user has scrolled to the bottom, to increase the clarity of the code. Also, to prevent the error, I gave the comparison a margin of error by testing that the value is below 1 instead of equal to 0.