Change the way clicking on a booklet sub-material is logged
Clicking on a sub-material in a booklet used to only be logged when the user clicked on the sub-material in the TOC or when they clicked the carousel arrows, not when swiping or using the pop-out TOC in the upper right corner. The function call for this logging was moved to the function handlePostChangeEventCarousel(), which is called by all ways of opening a booklet sub-material (TOC click, carousel arrows, swiping, pop-out TOC). This way, opening a sub-material is logged in all cases. There is only one exception: handlePostChangeEventCarousel() is not called when clicking on the first item in a TOC. This special case was handled separately.