Move the 'scrolled to bottom' test logic into a single function
The tests for whether the user has scrolled to the bottom of the screen appeared in four different places, causing code repetition. These four pieces of code were replaced by a function (logScrolledToBottomIfApplicable) and helper function (hasReachedBottomOfScreen) in educationalMaterialService, to improve clarity and code reuse. The two functions calling the scrolled to bottom requests (logScrolledToBottomEduMaterial and logSubScrolledToBottomEduMaterial) had to be moved from this service's export functions to its private functions.