Fix: Only log 'scrolled to bottom' on scrollable materials
The previous way of checking that a material was scrollable was unnecessarily costly (done from inside the scroll logging function). It was also buggy for booklets. A better way of checking scrolling on valid materials is to check the material type before activating the scroll and click listeners. This was implemented here. Now, scroll and click listeners are only activated for booklet sub-materials and for individual HTML pages.