Default to icon and colour of type 'Other' for educational materials of unknown types
Added a check before assigning an icon and a colour to an educational material based on its type. Now, if that type is not found in the list, the educational material will be assigned the icon and color or type 'Other' instead of nothing at all.
Also, removed the 'TestingType' icon and colour. Materials of type 'TestingType' will default to the 'Other' icon and colour from now on.