Issue 517: make lab results header sticky and remove duplicate html
Made the search bar and table header bar in the lab results test-view page sticky to lock them to the top of the screen while scrolling.
Removed duplicate code in the test-view html. Before, the code to generate each row was in triplicate (one copy for black, one for purple, one for red), with ng-if directives to determine which color to show. I removed the copied code and added functions in the controller to get the color class from there and bind it to the html with ng-class.
I also removed some duplicate or redundant classes in app.css, and fixed a French translation.