517: Add chevrons to lab results test-view
The lab results in the test view (test values for a single date) are clickable, but they did not have chevrons. I added chevrons to indicate at a glance that the tests can be clicked on to open the timeline-test-component view. To do this, I enabled the chevron modifier and added a fourth empty column in the grid to make space for the chevrons. To specify column width, I removed some bootstrap classes that were not working due to being incompatible with ons-row and ons-col, and used width values totaling 100% instead.