Rotate date labels on lab results graph
Rotated the date labels on the x-axis of the lab results graph to prevent them from overlapping. This issue was especially salient when the dates were in French. We might be able to return the labels to horizontal and have them not overlap when we update Highcharts, but there is no good way to do it in this version.