Fix bug with remembering lab graph timeframe
The function to save the user's choice of date range timeframe relied on checking the contents of the button label. For buttons where the label was different in English and French, the timeframe would not be saved for French. I fixed this bug by identifying each button by a number instead of by the button label.