412: Make response messages identical in Forgot Password
Made the generic responses messages identical when resetting your password. Now, the same message is shown whether the email is associated with a real Opal account or not. This protects against the case where a robot could try a list of email addresses and determine who is an Opal user and who is not.