Fix bug where rating edu material using the popover doesn't work
Using the table of contents popover to rate an educational booklet didn't work because the variable used to provide the EducationalMaterialControlSerNum was undefined. This resulted in all ratings having EducationalMaterialControlSerNum = 0. I changed the reference to the correct variable, so that the right EducationalMaterialControlSerNum appears in the database for each rating.