Removed duplicate code for questionnaire answer options
In commit f7e1ae3c98946e84a573102d483d2d41d075eba5, code for the questionnaire answer options was duplicated to allow the answers to be highlighted in blue when selected and white when unselected. This was done because AngularJS doesn't allow a variable in the HTML style tag. However, AngularJS does provide an ng-style tag that allows variables. I used this tag to remove the duplicate code. To do this, I added a new function 'styleIfSelected' in the controller to return the style of a button depending on whether or not it's selected.