Bug-fix: Multiple choice answers disappear when moving between questions in a questionnaire
When answering a multiple choice question in a questionnaire, the answers would get stored correctly to be sent to the database, but they would disappear from the screen when moving between questions. I fixed the function that loaded up the answers to be displayed, so the answers don't disappear anymore. I also commented out the part of this function dealing with 'other' answers, because this functionality was removed (commented out elsewhere in the code).