Add Sample Size Information to Appointment Delays and Fix First Click Not Firing in Personal Waiting Time
Appointment Delays Controller now sends the size of each set on top
of its percentage, which is received as the z coordinate of the series
and added to the tooltip format. Waiting Time Service has a work around
now. The first click wasn't firing because the k-d tree is built
asynchronously. Setting kdNow to true manually forces it to be built
synchronously