hot fix: adding debuggable back into config.xml for android (#764)
Note that even in cordova --release mode, the debuggable="false" option would not be inserted into the AndroidManifest.xml if the option is not present in the config.xml. Therefore we have to add the debuggable="false" option back into the config.xml files for prod, preprod and staging env.