Remove the chevron icon from the doctor location and it's not tappable. (#759)
Issue number: #497. This issue is on the individual doctor's page inside my treating team option. When user tap on any option inside my treating team, It opens the page with doctor location email and phone number and each three options has the chevron icon which indicates that option is tappable. By keeping that icon in the location it makes confusing to the user because it doesn't open the location on map and it was never implemented. I removed that icon and location option is not tappable anymore.