QSCCD-86: Filter out questionnaires locked by other users (!948)
This MR makes the remaining changes for questionnaire locking in the app: filtering out in-progress questionnaires that are locked by other users. I implemented this in the app only, because the app has to do this filtering anyways when it discovers that a questionnaire has been locked by another user. Because of this, it would be redundant to do the filtering in the listener too.

  * Also made a fix to ensure that when questionnaires are refreshed and a questionnaire changes status (e.g. in progress -> completed), the questionnaire is moved to the correct tab and the old copy is not left behind in the other tab. This is important when another user completes a questionnaire on another device, so that it displays correctly for the first user who didn't answer it.

See merge request opalmedapps/opal-app#948