ci: automatically set the app version via CI/CD pipeline
Installed and configured `semantic-release` and its plugins to set a new version number. Versions are tracked using tags. The biggest addition to the plugins' default settings was to execute our own script to change the version in `config.xml`. This is needed so that the version number gets used by Cordova and appears in the build.

Also refactored the pipeline config file to pull out common code (using `extends` and YAML anchors).

See merge request opalmedapps/qplus!988