fix: remove unused Token request parameter
This MR removes the Token parameter as a required field for request objects sent to the listener. This parameter was never directly used or validated (only saved to the database), but its value contained sensitive data (the Firebase user's JSON Web Token), so it was decided for security reasons that this token should be removed if not used.

*Additional*
- Deleted an old file containing firebase rules. These rules are now tracked and version-controlled in the `opal-listener` project.

See merge request opalmedapps/qplus!996