fix: set android:usesCleartextTraffic back to false for security purposes
Reset `android:usesCleartextTraffic` back to false. This setting was changed in !851 to allow access to the French version of labtestsonline (http://www.labtestsonline.fr/), which is now no longer available. Changed the setting back to its secure value (false), as recommended in the 2020 pen test.

See merge request opalmedapps/qplus!1052