fix: download full questionnaire stub via notifications instead of partial data
Fixed a bug where resuming a questionnaire via its notification would fail. This was due to the fact that the questionnaires notification redirect would download partial data (not a full notification stub), and would not save this data to the notification service. Fixed this issue by downloading and saving the full notification stub -- the same content that is downloaded when visiting the Questionnaires list.

Removed the previous call to `QuestionnaireInOpalDBFromSerNum`, and marked that request as deprecated in the listener.

See merge request opalmedapps/opal-app#1062