fix: added cancellable promise to fix login bug
Fixed a bug occurring at login when accidentally picking an invalid hospital and quickly switching to a valid one; the timeout from the first request would be handled mid-login and log the user out of Firebase.

Fixed this by adding the ability to cancel a request to the listener. When cancelled, the request still executes on the listener (there would be no way to recall it, though we don't need to do this here), but RequestToServer no longer processes the response, and the `.then()` of the request promise isn't triggered. Instead, an error of a specific type is rejected for `.catch()` to process as a special case.

Used this on the login page to cancel in-progress requests when leaving the page. Also made sure to apply this fix to all methods of leaving the page: going to the hospital selector, going to "Forgot password?", and clicking on the Back button.

**Other**

* Deleted unused `goToInit` function.
* Deleted `haveBeenSet` and `haveBeenSent` variables related to device identifiers, which were not being used productively and were causing issues.
* Disabled the trusted toggle during login.

See merge request opalmedapps/qplus!1071