chore: add required NSLocationAlwaysAndWhenInUseUsageDescription for App Store
Added the `NSLocationAlwaysAndWhenInUseUsageDescription` string required by the App Store in an email. Even if we don't use the location "always", the App Store flagged this feature as being part of one of our plugins (likely `cordova-plugin-geolocation`), so it requires a description.

After adding this string and re-uploading, we didn't get a new email about the issue.

See merge request opalmedapps/opal-app#1116