fix: adjust the space allocation in the upper toolbar
Adjusted the width allocated to each slot in the upper toolbar to leave more space for the page title, without affecting content in the left and right slots.

Note that the left and right content slots need to have the same width for the title to be centered.

See merge request opalmedapps/qplus!1181