fix: update hospital list
Updated the hospital list at login for open source. Removed the MUHC and CHU-SJ in preprod/prod. Now, the options are OMI and OHIGPH in our development-level environments, and just RI-MUHC in preprod/prod.

Also disabled the Cedars and MUHC links for the RI. We can decide later whether we want to completely remove these.

See merge request opalmedapps/opal-app#1234