fix: remove committed `env` files from the repository
Removed the contents of the `env` folder from the repository, leaving only instructions on how to configure your own environment. This was done to make things cleaner for open-source.

Relevant updates to the README file (regarding the installation process and how it relates to `env`) have been moved to another MR.

See merge request opalmedapps/qplus!1235