refactor: remove unused patient.png
Removed the `patient.png` image which is now only used as a fallback if the Home tab initialization code fails to execute correctly. Replaced this with a simple white circle as the fallback instead (since the user shouldn't ever see it, it doesn't have to be detailed).

See merge request opalmedapps/opal-app#1344