ci: add jobs to build and release the app to the app stores (#1460)
Added new jobs to our `Build and Deploy App` workflow to build release versions of the app and upload them to the stores. Store upload is done automatically for iOS and manually (for now) for Android.

Also updated our documentation on pipeline builds, and added new `Quick Guide` and `Pipeline Maintenance` sections.

Co-authored-by: Matthias Schoettle <git@mattsch.com>